Revert "feature/mdrp_28_rctva (#93)"
This reverts commit 058a9bc1930147da93c833b06be69250373d46c6.
