+++ config.py
@@ -1,5 +1,5 @@
 port = 5000
-workers = 1
+workers = 12
 bind = '0.0.0.0:{}'.format(port)
 country_package = 'openfisca_pf'
 timeout = 3600
