+++ openfisca_pf/base.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 
 # This file defines the base enum needed by our legislation.
-import numpy
 from openfisca_core.model_api import *
+import numpy
 from openfisca_core.taxscales import MarginalRateTaxScale
