+++ openfisca_pf/parameters/dicp/cps.yaml
@@ -1,0 +1,8 @@
+description: Taux de la CPS
+metadata:
+  reference: https://www.impot-polynesie.gov.pf/code/5-titre-v-contribution-pour-la-solidarite
+taux:
+  description: Taux de CPS
+  values:
+    2000-11-01: 0
+    2022-04-01: 0.01
