--- openfisca_pf/parameters/dicp/cps/taux.yaml
@@ -1,5 +1,0 @@
-description: Taux de la CPS
-reference: https://www.impot-polynesie.gov.pf/code/5-titre-v-contribution-pour-la-solidarite
-values:
-  2000-11-01: 0
-  2022-04-01: 0.01
