--- openfisca_pf/parameters/dicp/tva/seuils/credit_tva.yaml
@@ -1,57 +1,0 @@
-remboursement_possible_par_mois:
-  description: Mois de déclaration de TVA quand il est possible de demander un remboursement de crédit de TVA.
-  metadata:
-    reference: https://www.impot-polynesie.gov.pf/code/4-section-iv-remboursement-de-credit-de-taxe-deductible
-  "01":
-    1995-02-02: false
-  "02":
-    1995-02-02: false
-  "03":
-    1995-02-02: true
-  "04":
-    1995-02-02: false
-  "05":
-    1995-02-02: false
-  "06":
-    1995-02-02: true
-  "07":
-    1995-02-02: false
-  "08":
-    1995-02-02: false
-  "09":
-    1995-02-02: true
-  "10":
-    1995-02-02: false
-  "11":
-    1995-02-02: false
-  "12":
-    1995-02-02: true
-
-minimum_par_mois:
-  description: Montant requis au mois de la déclaration pour prétendre à un remboursement de crédit de TVA.
-  metadata:
-    reference: https://www.impot-polynesie.gov.pf/code/4-section-iv-remboursement-de-credit-de-taxe-deductible
-  "01":
-    1995-02-02: 1.7976931348623157e+308
-  "02":
-    1995-02-02: 1.7976931348623157e+308
-  "03":
-    1995-02-02: 100_000
-  "04":
-    1995-02-02: 1.7976931348623157e+308
-  "05":
-    1995-02-02: 1.7976931348623157e+308
-  "06":
-    1995-02-02: 100_000
-  "07":
-    1995-02-02: 1.7976931348623157e+308
-  "08":
-    1995-02-02: 1.7976931348623157e+308
-  "09":
-    1995-02-02: 100_000
-  "10":
-    1995-02-02: 1.7976931348623157e+308
-  "11":
-    1995-02-02: 1.7976931348623157e+308
-  "12":
-    1995-02-02: 20_000
