--- openfisca_pf/tests/dicp/tva/tests_credit_tva.yaml
@@ -1,38 +1,0 @@
-- name: 1
-  period: 2023-06
-  input:
-    base_imposable_tva_taux_reduit: 1000000
-    base_imposable_tva_taux_intermediaire: 1000000
-    base_imposable_tva_taux_normal: 1000000
-    regularisation_autre_tva_exigible: 0
-    montant_ventes_hors_taxes: 1500000
-    montant_prestations_services_hors_taxes: 1500000
-    tva_immobilisation_deductible: 0
-    tva_autres_biens_deductibles: 0
-    regularisation_tva_deductible: 0
-    report_credit_tva_deductible: 0
-    rib_renseigne: true
-    cps_en_diminution: 0
-    cps_a_reverser: 0
-  output:
-    entrants_tva_valides: true
-    tva_due_taux_reduit: 50000.0
-    tva_due_taux_intermediaire: 130000.0
-    tva_due_taux_normal: 160000.0
-    tva_exigible: 340000.0
-    tva_deductible: 0
-    tva_nette_due: 340000.0
-    credit_tva: 0.0
-    credit_tva_minimum_ce_mois: 100000.0
-    remboursement_credit_tva_possible_car_rib_rensigne: true
-    remboursement_credit_tva_possible_ce_mois: true
-    remboursement_credit_tva_possible_car_montant_suffisant: false
-    remboursement_credit_tva_possible: false
-    demande_rembousement_credit_tva_valide: false
-    remboursement_credit_tva: 0
-    credit_tva_a_reporter: 0
-    base_imposable_cps: 3000000.0
-    cps_due: 30000.0
-    cps_nette: 30000.0
-    cps_nette_due: 30000.0
-    cps_a_reporter: 0.0
