+++ openfisca_pf/variables/dicp/TVA/tva_pays.py
@@ -1,6 +1,12 @@
 # -*- coding: utf-8 -*-
 
-from openfisca_pf.constants import units
+# This file defines variables for the modelled legislation.
+# A variable is a property of an Entity such as a Person, a Household…
+# See https://openfisca.org/doc/key-concepts/variables.html
+
+# Import from openfisca-core the common Python objects used to code the legislation in OpenFisca
+from openfisca_core.model_api import *
+# Import the Entities specifically defined for this tax and benefit system
 from openfisca_pf.entities import *
 from openfisca_pf.base import *
 
@@ -10,7 +16,6 @@ class tva_nette_due_total_pays(Variable):
     entity = Pays
     definition_period = MONTH
     label = u"Montant total de TVA nette due par les entreprises du pays"
-    unit = units.XPF
 
     def formula(pays, period, parameters):
         tva_nette_due = pays.members('tva_nette_due', period)
@@ -22,7 +27,6 @@ class credit_tva_total_pays(Variable):
     entity = Pays
     definition_period = MONTH
     label = u"Montant total de crédit de TVA à à rembourser aux entreprises du pays"
-    unit = units.XPF
 
     def formula(pays, period, parameters):
         credit_tva = pays.members('credit_tva', period)
@@ -34,7 +38,6 @@ class tva_due_taux_reduit_pays(Variable):
     entity = Pays
     definition_period = MONTH
     label = u"Montant total de TVA due en taux réduit par les entreprises du pays"
-    unit = units.XPF
 
     def formula(pays, period, parameters):
         tva_due_taux_reduit_pays = pays.members('tva_due_taux_reduit', period)
@@ -46,7 +49,6 @@ class tva_due_taux_intermediaire_pays(Variable):
     entity = Pays
     definition_period = MONTH
     label = u"Montant total de TVA due en taux intermediaire par les entreprises du pays"
-    unit = units.XPF
 
     def formula(pays, period, parameters):
         tva_due_taux_intermediaire_pays = pays.members('tva_due_taux_intermediaire', period)
@@ -58,7 +60,6 @@ class tva_due_taux_normal_pays(Variable):
     entity = Pays
     definition_period = MONTH
     label = u"Montant total de TVA due en taux normal par les entreprises du pays"
-    unit = units.XPF
 
     def formula(pays, period, parameters):
         tva_due_taux_normal_pays = pays.members('tva_due_taux_normal', period)
@@ -70,7 +71,6 @@ class tva_nette_due_total_pays_annee(Variable):
     entity = Pays
     definition_period = YEAR
     label = u"Montant total de TVA nette due par les entreprises du pays sur l'année"
-    unit = units.XPF
 
     def formula(pays, period, parameters):
         tva_nette_due_total_pays_annee = pays('tva_nette_due_total_pays', period, options = [ADD])
@@ -82,7 +82,6 @@ class credit_tva_total_pays_annee(Variable):
     entity = Pays
     definition_period = YEAR
     label = u"Montant total de crédit de TVA à à rembourser aux entreprises du pays sur l'année"
-    unit = units.XPF
 
     def formula(pays, period, parameters):
         credit_tva_total_pays_annee = pays('credit_tva_total_pays', period, options = [ADD])
@@ -94,7 +93,6 @@ class tva_due_taux_reduit_pays_annee(Variable):
     entity = Pays
     definition_period = YEAR
     label = u"Montant total de TVA due en taux réduit par les entreprises du pays sur l'année"
-    unit = units.XPF
 
     def formula(pays, period, parameters):
         tva_due_taux_reduit_pays_annee = pays('tva_due_taux_reduit_pays', period, options = [ADD])
@@ -106,7 +104,6 @@ class tva_due_taux_intermediaire_pays_annee(Variable):
     entity = Pays
     definition_period = YEAR
     label = u"Montant total de TVA due en taux intermediaire par les entreprises du pays sur l'année"
-    unit = units.XPF
 
     def formula(pays, period, parameters):
         tva_due_taux_intermediaire_pays_annee = pays('tva_due_taux_intermediaire_pays', period, options = [ADD])
@@ -118,7 +115,6 @@ class tva_due_taux_normal_pays_annee(Variable):
     entity = Pays
     definition_period = YEAR
     label = u"Montant total de TVA due en taux normal par les entreprises du pays sur l'année"
-    unit = units.XPF
 
     def formula(pays, period, parameters):
         tva_due_taux_normal_pays_annee = pays('tva_due_taux_normal_pays', period, options = [ADD])
