+++ setup.py
@@ -4,7 +4,7 @@ from setuptools import setup, find_packages
 
 setup(
     name = "OpenFisca-PF",
-    version = "0.6.3",
+    version = "0.6.2",
     author = "SIPf",
     author_email = "matthieu.bosc@informatique.gov.pf",
     classifiers = [
