+++ start.sh
@@ -1,1 +1,2 @@
+killall openfisca
 openfisca serve --configuration-file config.py
